fixed #3

+++ src/Bar/Bar.cpp
@@ -8,15 +8,15 @@ Bar::Bar(BarState newBarState)
 	mySkin = LoadTexture("resource/assets/img/Kenney.nl/png/bar.png");
 	barStatus = newBarState;
 	remberFirstStatus = newBarState;
-	mySkin.width = ratioReductionTexture * 3;//mySkin.width / ratioReductionTexture;
-	mySkin.height = ratioReductionTexture;//mySkin.height / ratioReductionTexture;
+	mySkin.width = static_cast<int>(ratioReductionTexture * 3);//mySkin.width / ratioReductionTexture;
+	mySkin.height = static_cast<int>(ratioReductionTexture);//mySkin.height / ratioReductionTexture;
 	position.x = 0;
 	position.y = 0;
 }
 void Bar::setPosition(int x, int y)
 {
-	position.x = x;
-	position.y = y;
+	position.x = static_cast<float>(x);
+	position.y = static_cast<float>(y);
 }
 void Bar::setStatus(BarState newStatus)
 {
@@ -45,7 +45,7 @@ void Bar::showMe()
 {
 	if (barStatus == VISIBLE)
 	{
-		DrawTexture(mySkin, position.x, position.y, WHITE);
+		DrawTexture(mySkin, static_cast<int>(position.x), static_cast<int>(position.y), WHITE);
 	}
 }
 int Bar::getDistanceBallIncideBar()
